Move the ls functionality to separate file
